requirements.txt
discord.py==2.1.0
python-dotenv==0.21.1
asgiref==3.6.0
openai==0.27.0
pinecone-client==2.2.1